+++ src/main/java/com/shaoxing/lixing/domain/entity/FCompanyBill.java
@@ -66,7 +66,7 @@ public class FCompanyBill extends Model<FCompanyBill> {
     /**
      * 数量
      */
-    private Integer num;
+    private BigDecimal num;
 
     /**
      * 重量
